+++ spec/download_spec.rb
@@ -15,6 +15,11 @@ RSpec.describe OnlyofficeS3Wrapper do
     expect(File).to exist("#{s3.download_folder}/Book.docx")
   end
 
+  it 'download_file_by_name return result file path' do
+    file = s3.download_file_by_name('docx/Book.docx')
+    expect(File).to exist(file)
+  end
+
   it 'download_object for nonexitsing name cause exception' do
     fake_name = 'docx/fake-name.fakeext'
     expect { s3.download_file_by_name(fake_name) }.to raise_error(/not found/)